third_party/WebKit/LayoutTests/html/tabular_data/col_span.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("HTMLTableColElement span attribute test");

function spanAttributeEffect(value)
{
    var element = document.createElement("col");
    element.setAttribute("span", value);
    return element.span;
}

shouldBe('document.createElement("col").span', '1');

shouldBe('spanAttributeEffect("")', '1');

shouldBe('spanAttributeEffect("1")', '1');
shouldBe('spanAttributeEffect("2")', '2');
shouldBe('spanAttributeEffect("10")', '10');

shouldBe('spanAttributeEffect("0")', '1');

shouldBe('spanAttributeEffect("-1")', '1');

shouldBe('spanAttributeEffect("1x")', '1');
shouldBe('spanAttributeEffect("1.")', '1');
shouldBe('spanAttributeEffect("1.9")', '1');
shouldBe('spanAttributeEffect("2x")', '2');
shouldBe('spanAttributeEffect("2.")', '2');
shouldBe('spanAttributeEffect("2.9")', '2');

shouldBe('spanAttributeEffect("a")', '1');
shouldBe('spanAttributeEffect("\v7")', '1');
shouldBe('spanAttributeEffect("  7")', '7');

var arabicIndicDigitOne = String.fromCharCode(0x661);
shouldBe('spanAttributeEffect(arabicIndicDigitOne)', '1');
shouldBe('spanAttributeEffect("2" + arabicIndicDigitOne)', '2');

shouldBe('spanAttributeEffect("2147483647")', '1000');
shouldBe('spanAttributeEffect("2147483648")', '1000');
shouldBe('spanAttributeEffect("4294967295")', '1000');
shouldBe('spanAttributeEffect("4294967296")', '1000');
</script>
</body>
</html>
