third_party/blink/web_tests/platform/generic/html/tabular_data/col_span-expected.txt

HTMLTableColElement span attribute test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.createElement("col").span is 1
PASS spanAttributeEffect("") is 1
PASS spanAttributeEffect("1") is 1
PASS spanAttributeEffect("2") is 2
PASS spanAttributeEffect("10") is 10
PASS spanAttributeEffect("0") is 1
PASS spanAttributeEffect("-1") is 1
PASS spanAttributeEffect("1x") is 1
PASS spanAttributeEffect("1.") is 1
PASS spanAttributeEffect("1.9") is 1
PASS spanAttributeEffect("2x") is 2
PASS spanAttributeEffect("2.") is 2
PASS spanAttributeEffect("2.9") is 2
PASS spanAttributeEffect("a") is 1
PASS spanAttributeEffect("7") is 1
PASS spanAttributeEffect("  7") is 7
PASS spanAttributeEffect(arabicIndicDigitOne) is 1
PASS spanAttributeEffect("2" + arabicIndicDigitOne) is 2
PASS spanAttributeEffect("2147483647") is 1000
PASS spanAttributeEffect("2147483648") is 1000
PASS spanAttributeEffect("4294967295") is 1000
PASS spanAttributeEffect("4294967296") is 1000
PASS successfullyParsed is true

TEST COMPLETE
